added catalog view, actions, service methods

+++ src/main/java/com/java/cruisecompany/model/dto/RouteDTO.java
@@ -4,12 +4,14 @@ import com.java.cruisecompany.model.entity.Port;
 import com.java.cruisecompany.model.entity.wrapper.Waypoint;
 import lombok.Builder;
 import lombok.Data;
+import lombok.ToString;
 
 import java.math.BigDecimal;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Builder
 @Data
@@ -20,6 +22,6 @@ public class RouteDTO {
     private LocalDate endOfCruise;
     private BigDecimal price;
     private List<Waypoint> waypoints;
-    private long duration;
+    private int duration;
     private long numOfPorts;
 }
